build: Expand the list of directories for phan to analyze
The phan config only looks for includes/, src/ and maintenance/ folder
Just running for maintenance folder and also exclude it from analysis
makes it a no-operation

- Add undeclared class properties
- Declare arrays before use
- Add inline phan-var suggestion
- Add some casts for string/int
- Split array_shift on two lines as this needs by-ref
- throw from not implemented functions
- Fix param docs
- Suppress some issues inline
- Fix string interpolate with bool types
- Whitelist a raw html messages
- Remove unreachable code after new operator or not-null type-hints

Bug: T224763
Follow-Up: If1dda36d9ec39e1a9b55d86f4be1738d8a60617f

+++ gateway_common/AmountInCents.php
@@ -21,7 +21,7 @@ class AmountInCents implements StagingHelper, UnstagingHelper {
 			return;
 		}
 
-		$amount = Amount::round( $normalized['amount'], $normalized['currency'] );
+		$amount = (float)Amount::round( $normalized['amount'], $normalized['currency'] );
 
 		$stagedData['amount'] = round( $amount * 100 );
 	}
